tier5 phase 5: make std-collapse regression test real + std result dtype consistency (review Minors)
Minor 1 (critical guard): test_zero_std_large_constant_logits used logits [30,30,30,30]
where sigmoid saturates exactly to 1.0 in float32, making E[s²]-E[s]² = 1·1 - 1·1 = 0.0
exactly. This allowed the pre-fix (buggy float32-accumulation) code to pass—hollow guard.
Replaced with test_near_saturated_varied_logits_std_no_float32_noise using logits
[15,16,17] where sigmoid is near-but-not-exactly-constant. Pre-fix float32 accumulation
yields std ≈ 2.44e-4 (2^-12 cancellation error), oracle ≈ 1.5e-7. Tolerance abs=1e-5
fails pre-fix but passes float64 fix, deterministically guarding Phase 7/8 std≈0 collapse
signal against float64→float32 regression. Verified pre-fix fails: 2.441406e-04 > 1e-5.

Minor 2 (cosmetic, now fixed): PredictionStdMetric.result() returned float64 tensor while
PredictionMeanMetric/PredictionFracAboveMetric returned float32 (metric.dtype=float32).
Harmless but inconsistent. Cast final result to self.dtype, preserving float64 accumulation
and subtraction arithmetic for numerical precision while making return dtype consistent.

+++ src/diagnostics/distribution_metrics.py
@@ -105,7 +105,11 @@ class PredictionStdMetric(keras.metrics.Metric):
         mean = ops.divide_no_nan(self._sum, self._count)
         mean_sq = ops.divide_no_nan(self._sum_sq, self._count)
         var = ops.maximum(mean_sq - mean * mean, 0.0)
-        return ops.sqrt(var)
+        # Cast to self.dtype for consistency with sibling metrics (PredictionMeanMetric
+        # and PredictionFracAboveMetric return float32). Accumulation+subtraction
+        # arithmetic stays in float64 for numerical precision; only the final result
+        # is cast back to the metric's declared dtype.
+        return ops.cast(ops.sqrt(var), self.dtype)
 
     def reset_state(self):
         self._sum.assign(0.0)
